requirements.txt
# Kaori Protocol Dependencies

# Core API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
requests>=2.31.0

# Database
sqlalchemy>=2.0.0

# Data Validation & Auth
pydantic>=2.0.0
python-jose>=3.3.0

# AI & Computer Vision
transformers>=4.30.0
torch>=2.0.0
pillow>=10.0.0
numpy>=1.24.0

# Testing
pytest>=7.0.0
